backend/requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
geoalchemy2==0.14.3
psycopg2-binary==2.9.9

# Data processing
pandas==2.2.0
openpyxl==3.1.2
numpy==1.26.3

# Google Cloud
google-cloud-storage==2.14.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0

# HTTP client
httpx==0.26.0
aiohttp==3.9.3

# Task queue
celery==5.3.6
redis==5.0.1

# Twilio
twilio==8.12.0

# Utilities
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1
shapely==2.0.2
pyproj==3.6.1
aiofiles==23.2.1

# Testing
pytest>=7.0.0,<8.0.0
pytest-asyncio==0.23.4

